+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### 12.02.2018
 
+* [Rust] Bookshelf problem (Used shelves with books) (Rust)
 * [Rust] Bookshelf problem (Binary for challenge data testing) (Reddit)
 
 #### 11.02.2018